refactor(dependencies graph): moved to a dedicated file
